Actually, the load functionality wasnt implemented in the last commit, but in this one it is. Also, the concentration of a cds is now stored in the database as a list of variables

+++ app/models/Parts.scala
@@ -41,7 +41,7 @@ case class CodingSeq(name:String, var concentration: List[(Double, Double)]) ext
     /**
      * Save this codingSequence to the database.
      */
-    def save(id: Long) {
+    def save(id: Int) {
       DB.withConnection { implicit connection =>
           for(l <- linksTo){
 		      SQL("merge into cds key(id,name,next) values({id},{name},{next});")
